# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MDM Platform API
  description: REST API for the MDM admin console. Manages devices, policies, apps, compliance, and more.
  version: 1.0.0
  contact:
    name: MDM Platform Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.staging.mdm.example.com/api/v1
    description: Staging
  - url: https://api.mdm.example.com/api/v1
    description: Production
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    ApiError:
      type: object
      required: [statusCode, error, message]
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
        details:
          type: object
        requestId:
          type: string
...
    DeviceGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        isDynamic:
          type: boolean
        dynamicFilter:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateGroupRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        isDynamic:
          type: boolean
          default: false
        dynamicFilter:
          type: object
    UpdateGroupRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        dynamicFilter:
          type: object
    AddGroupMemberRequest:
      type: object
      required: [deviceId]
      properties:
        deviceId:
          type: string
          format: uuid
    GroupListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DeviceGroup'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # --- Command ---
...
    GroupMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NoContent:
      description: Successful deletion, no content returned
paths:
...
  /groups:
    get:
      operationId: listGroups
      summary: List device groups
      tags: [groups]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupListResponse'
    post:
      operationId: createGroup
      summary: Create a device group
      tags: [groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceGroup'
  /groups/{id}:
    get:
      operationId: getGroup
      summary: Get device group by ID
      tags: [groups]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceGroup'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateGroup
      summary: Update a device group
      tags: [groups]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceGroup'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteGroup
      summary: Delete a device group
      tags: [groups]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
  /groups/{id}/members:
    post:
      operationId: addGroupMember
      summary: Add a device to a group
      tags: [groups]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGroupMemberRequest'
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMember'
  /groups/{id}/members/{deviceId}:
    delete:
      operationId: removeGroupMember
      summary: Remove a device from a group
      tags: [groups]
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          $ref: '#/components/responses/NoContent'

  # ==================== COMMANDS ====================
...